Add files via upload
A unit category is selected by clicking/checking an appropriate radio button.
When a radio button is checked an appropriate text will be shown in the label (e.g. “Convert Length”, “Convert Weight”, etc.) and the corresponding unit’s options will be available in both combo boxes for the user to choose from.
In addition, all existing inputs/selections/results in both text boxes and combo boxes are deleted, and the focus is set on the first text box. After selecting one of three unit categories, the first panel (“From”) is used to enter a value into the text box and select an appropriate unit in the combo box. 
Then the second panel (“To”) to select the unit to which the previously entered value will be converted and the result of the unit’s conversion will be shown in the second text box. The “CLOSE” button will terminate the program.